+++ test/translations/nl.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'Hello' => 'Hallo',
+    'page' => 'pagina',
+    'order' => 'sortering',
+];